# hansenlaw vectorization with np.power()