--- a/examples/example_all_O2.py
+++ b/examples/example_all_O2.py
@@ -23,11 +23,13 @@ from time import time
 #   dictionary of method: function()
 
 transforms = {
-  "direct": abel.direct.direct_transform,      
-  "onion": abel.dasch.onion_peeling_transform, 
-  "hansenlaw": abel.hansenlaw.hansenlaw_transform,
   "basex": abel.basex.basex_transform,   
+  "direct": abel.direct.direct_transform,      
+  "hansenlaw": abel.hansenlaw.hansenlaw_transform,
+  "onion_bordas": abel.onion_bordas.onion_bordas_transform,
+  "onion_dasch": abel.dasch.onion_peeling_transform, 
   "three_point": abel.dasch.three_point_transform,
+  "two_point" : abel.dasch.two_point_transform,
 }
 # sort dictionary 
 transforms = collections.OrderedDict(sorted(transforms.items()))
@@ -36,28 +38,30 @@ ntrans = np.size(transforms.keys())  # number of transforms
 
 # Image:   O2- VMI 1024x1024 pixel ------------------
 IM = np.loadtxt('data/O2-ANU1024.txt.bz2')
-# recenter the image to an odd size
 
+# recenter the image to mid-pixel (odd image width)
 IModd = abel.tools.center.center_image(IM, center="slice", odd_size=True)
-# np.savetxt("O2-ANU1023.txt", IModd)
 
 h, w = IModd.shape
 print("centered image 'data/O2-ANU2048.txt' shape = {:d}x{:d}".format(h, w))
 
-Q0, Q1, Q2, Q3 = abel.tools.symmetry.get_image_quadrants(IModd, reorient=True)
+# split image into quadrants
+Q = abel.tools.symmetry.get_image_quadrants(IModd, reorient=True)
 
+Q0 = Q[0]
 Q0fresh = Q0.copy()    # keep clean copy
 print ("quadrant shape {}".format(Q0.shape))
 
 # Intensity mask used for intensity normalization
 #   quadrant image region of bright pixels 
-
 mask = np.zeros(Q0.shape, dtype=bool)
 mask[500:512, 358:365] = True   
 
 # process Q0 quadrant using each method --------------------
 
 iabelQ = []  # keep inverse Abel transformed image
+sp = []  # speed distributions
+meth = []  # methods
 
 for q, method in enumerate(transforms.keys()):
 
@@ -67,60 +71,65 @@ for q, method in enumerate(transforms.keys()):
     t0 = time()
 
     # inverse Abel transform using 'method'
-    IAQ0 = transforms[method](Q0, direction="inverse") 
-
+    IAQ0 = transforms[method](Q0, direction="inverse", dr=0.5) 
     print ("                    {:.1f} sec".format(time()-t0))
 
-    iabelQ.append(IAQ0)  # store for plot
 
     # polar projection and speed profile
-    radial, speed = abel.tools.vmi.angular_integration(IAQ0, origin=(0, 0))
+    radial, speed = abel.tools.vmi.angular_integration(IAQ0, origin=(0, 0), dr=0.5)
 
     # normalize image intensity and speed distribution
     IAQ0 /= IAQ0[mask].max()  
     speed /= speed[radial > 50].max()
 
-    # plots    #121 whole image,   #122 speed distributions
-    plt.subplot(121) 
-
-    # method label for each quadrant
-    annot_angle = -(45+q*90)*np.pi/180  # -ve because numpy coords from top
-    if q > 3: 
-        annot_angle += 50*np.pi/180    # shared quadrant - move the label  
-    annot_coord = (h/2+(h*0.9)*np.cos(annot_angle)/2, 
-                   w/2+(w*0.9)*np.sin(annot_angle)/2)
-    plt.annotate(method, annot_coord, color="yellow")
-
-    # plot speed distribution
-    plt.subplot(122) 
-    plt.plot(radial, speed, label=method)
+    # keep data for plots
+    iabelQ.append(IAQ0)  
+    sp.append((radial, speed))
+    meth.append(method)
 
 # reassemble image, each quadrant a different method
 
+# plot inverse Abel transformed image slices, and respective speed distributions
+ax0 = plt.subplot2grid((1, 2), (0, 0))
+ax1 = plt.subplot2grid((1, 2), (0, 1))
+
+def ann_plt (quad, subquad, txt):
+    # -ve because numpy coords from top
+    annot_angle = -(30+30*subquad+quad*90)*np.pi/180
+    annot_coord = (h/2+(h*0.8)*np.cos(annot_angle)/2,
+                   w/2+(w*0.8)*np.sin(annot_angle)/2)
+    ax0.annotate(txt, annot_coord, color="yellow", horizontalalignment='left')
+
 # for < 4 images pad using a blank quadrant
-blank = np.zeros(IAQ0.shape)  
-for q in range(ntrans, 4):
-    iabelQ.append(blank)
+r, c = Q0.shape
+Q = np.zeros((4, r, c))
 
-# more than 4, split quadrant
-if ntrans == 5:
-    # split last quadrant into 2 = upper and lower triangles
-    tmp_img = np.tril(np.flipud(iabelQ[-2])) +\
-              np.triu(np.flipud(iabelQ[-1]))
-    iabelQ[3] = np.flipud(tmp_img)
-# Fix me when > 5 images
- 
+indx = np.triu_indices(iabelQ[0].shape[0])
+iq = 0
+for q in range(4):
+    Q[q] = iabelQ[iq].copy()
+    ann_plt(q, 0, meth[iq])
+    ax1.plot(*(sp[iq]), label=meth[iq])
+    iq += 1
+    if iq < len(transforms):
+        Q[q][indx] = np.triu(iabelQ[iq])[indx] 
+        ann_plt(q, 1, meth[iq])
+        ax1.plot(*(sp[iq]), label=meth[iq])
+    iq += 1
 
-im = abel.tools.symmetry.put_image_quadrants((iabelQ[0], iabelQ[1],
-                                              iabelQ[2], iabelQ[3]), 
+# reassemble image from transformed (part-)quadrants
+im = abel.tools.symmetry.put_image_quadrants((Q[0], Q[1], Q[2], Q[3]), 
                                               original_image_shape=IModd.shape)
 
-plt.subplot(121)
-plt.imshow(im, vmin=0, vmax=0.8)
+ax0.axis('off')
+ax0.set_title("inverse Abel transforms")
+ax0.imshow(im, vmin=0, vmax=0.8)
 
-plt.subplot(122)
-plt.axis(ymin=-0.05, ymax=1.1, xmin=50, xmax=450)
-plt.legend(loc=0, labelspacing=0.1)
+ax1.set_title("speed distribution")
+ax1.axis(ymin=-0.05, ymax=1.1, xmin=50, xmax=450)
+ax1.legend(loc=0, labelspacing=0.1, fontsize=10)
 plt.tight_layout()
+
+# save a copy of the plot
 plt.savefig('example_all_O2.png', dpi=100)
 plt.show()