Using the magic of `np.power()` to replace `pow()` allows some vectorization improvement to the `hansenlaw` algorithm.  I imagine/hope these types of code changes may be taken further ...

- removes 'k' for loop

```python
for k in range(K):  # Iterate over k, the eigenvectors?
            X[:, k] = pow(Nm, lam[k])*X[:, k] +\
                      h[k]*gamma(Nm, lam[k], n)*gp[:, n]  # Eq. (15 or 17)
```
- (apparently) no longer requires the pseudo 1-column shift - e.g. run `examples/example_all_O2.py`

 ```python 
  #for some reason shift by 1 pixel aligns better? - FIX ME!
      if direction == "inverse":
          AIM = np.c_[AIM[:, 1:],AIM[:, -1]]
 ```
- faster execution  ~30% improvement (`n=501`) cf [PR #155](https://github.com/PyAbel/PyAbel/pull/155#issuecomment-203097576)
```
In [2]: abel.benchmark.AbelTiming(n=[301, 501, 1001])
Out[2]: 
PyAbel benchmark run on x86_64

time in milliseconds

=========    Basis Abel implementations ==========

Implementation      n = 301           n = 501           n = 1001      
----------------------------------------------------------------------------
        basex_bs    6148.8           20052.1               nan         
onion_peeling_bs       6.5              26.3             172.5         
  three_point_bs      10.8              28.4             135.1         
    two_point_bs       4.3              11.8              43.6         


=========  Forward Abel implementations ==========

Implementation      n = 301           n = 501           n = 1001      
----------------------------------------------------------------------------
   direct_Python      65.0             297.8               nan         
       hansenlaw       9.8              21.6              75.6         


=========  Inverse Abel implementations ==========

Implementation      n = 301           n = 501           n = 1001      
----------------------------------------------------------------------------
           basex      15.0             120.1               nan         
   direct_Python      74.8             313.0               nan         
       hansenlaw      10.1              21.6              71.6         
    onion_bordas     205.2             751.2            3187.9         
   onion_peeling       4.2              18.8             140.6         
     three_point       4.7              19.3             142.1         
       two_point       4.7              19.4             143.7    
```
![pyabel-benchmarks](https://cloud.githubusercontent.com/assets/10932229/14266254/3dc3243e-fb0a-11e5-847e-dbf488d24f92.png)