@DanHickstein 2016-04-04
@stggh, this looks like a nice improvement!

Can you make a comparison of both hansenlaw methods out to some large image size? (Is 5000x5000 feasible?)

I assume that this aggressively-vectorized hansenlaw will surely be faster for small images, but we should just make sure that it also works for larger images as well. (It will take more RAM, so this may be a problem for super-huge images. I'm just not sure what "super huge" is in this case.)

(Those benchmarks are from the Ubuntu computer with the slow Numpy, right? The Dasch methods look a little bit slow, and this is probably a result of us switching them to `np.tensordot`, which doesn't run quickly when linked with non-optimized linear-algebra libraries.)