@stggh 2016-04-04
Yes, I had also wondered about the image sizes effect.  Also, the `direct_C` is missing from the above benchmarks, for some reason it goes "missing" on my Ubuntu system. 

New `hansenlaw` benchmarks for large `n` images:
 `n<3001`
```
Implementation      n = 501           n = 1001          n = 2001          n = 3001      
----------------------------------------------------------------------------------------------
       hansenlaw      20.1              73.5             262.7             570.9         
   onion_peeling      18.5             138.5            1099.1            3614.5         
     three_point      18.6             139.4            1083.9            3637.9         
       two_point      18.7             143.1            1099.4            3649.3 
```
 `n<10001`  (only `hansenlaw` has a sensible execution time):
```
Implementation      n = 5001          n = 8001          n = 10001     
----------------------------------------------------------------------------
       hansenlaw    1764.2            4079.1            6491.0        
```

Old `hansenlaw`:
```
Implementation      n = 1001          n = 2001          n = 3001          n = 5001      
----------------------------------------------------------------------------------------------
       hansenlaw      87.2             307.0             663.5            2006.4 
```
```
=========  Inverse Abel implementations ==========

Implementation      n = 5001          n = 8001          n = 10001     
----------------------------------------------------------------------------
       hansenlaw    1827.5            4583.2            7038.6      
```

The relative execution time improvement for larger images is less.