@DanHickstein 2016-04-04
Awesome! Thanks for doing that. Pretty much as expected. At some huge image size, "old" hansenlaw might be faster than new Hansenlaw. But, I think that we can optimize for images less than 10,000x10,000 :) 

The improvement with the new Hansenlaw is great in the ~1000x1000 range.