@stggh 2016-04-06
I have not had much free time to play, but moving the k-vector loop outside of the column iteration slightly improves on the execution times from those above.  My gut feeling is that he `i` column loop may eventually be eliminated. Core `hansenlaw.py` code:
```python
    # special case lam = 0
    hkgam[:, 0] = h[0]*gammagt(Nm, lam[0], nn)
    pNm[:, 0] = np.ones_like(Nm)

    # lam < 0
    for k in range(1, K):
        hkgam[:, k] = h[k]*gammalt(Nm, lam[k], nn)
        pNm[:, k] = np.power(Nm, lam[k])

    # Eq. (15) or (17)
    for i, n in enumerate(nn):
        X = pNm[i]*X + hkgam[i]*np.expand_dims(gp[:, n], axis=1)
        AIM[:, n] = X.sum(axis=1)
```

```
=========  Inverse Abel implementations ==========

Implementation      n = 501           n = 1001          n = 2001          n = 3001      
----------------------------------------------------------------------------------------------
       hansenlaw      18.3              64.4             248.2             557.8         
   onion_peeling      19.0             142.1            1097.0            3592.8         
     three_point      19.3             140.3            1084.3            3864.9         
       two_point      19.3             139.1            1088.3            3786.6    
```
```
Implementation      n = 5001          n = 8001          n = 10001     
----------------------------------------------------------------------------
       hansenlaw    1591.8            4009.1            5975.3         
```
![pyabel-benchmarks](https://cloud.githubusercontent.com/assets/10932229/14334847/7559f986-fc9a-11e5-8e54-f2564680eb2d.png)