@DanHickstein 2016-04-07
This is interesting - it looks like you are pre-computing something that looks like a basis set, and then you are applying this to the image, in a way that is independent of the row and summing over the columns. It seems to me that you are about one step away from re-phrasing the Hansen-Law transform in terms of a dot product, just like the other transforms. 

It's interesting how this algorithm still scales much better than the other transforms as we go to huge image sizes. 

You can feel free to merge this if you want. Or keep playing with it.