@stggh 2016-04-07
I'll keep playing with it over the weekend. Yes, execution time improvements, ultimately leading to the dot product, is the goal ;-) 

@rth it is great to hear from you. PyAbel misses your clever python coding and in-depth discussion.